Better error handling in many places, refactorings, more tests, nicer interface go obtain clients

+++ common_fields.go
@@ -5,8 +5,8 @@ import (
 )
 
 type CommonFields struct {
-	ID        int       `json:"id"`
-	ShopID    int       `json:"shop_id"`
+	ID        int64     `json:"id"`
+	ShopID    int64     `json:"shop_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
